+++ src/multipliers.lua
@@ -8,8 +8,8 @@ game.TraitData.AphroditeSpecialBoon.RarityLevels.Perfect = { Multiplier = 3.5 } 
 game.TraitData.AphroditeCastBoon.RarityLevels.Perfect = { Multiplier = 3.5 }
 game.TraitData.AphroditeSprintBoon.RarityLevels.Perfect = { Multiplier = 3.5 }
 game.TraitData.AphroditeManaBoon.RarityLevels.Perfect = { Multiplier = 2.67 } -- 0.33
-game.TraitData.HighHealthOffenseBoon.RarityLevels.Perfect = { Multiplier = 20/10 } -- updated 2/19
-game.TraitData.HealthRewardBonusBoon.RarityLevels.Perfect = { Multiplier = 2.25 }
+game.TraitData.HighHealthOffenseBoon.RarityLevels.Perfect = { Multiplier = 15 / 5 } -- updated 6/17
+game.TraitData.HealthRewardBonusBoon.RarityLevels.Perfect = { Multiplier = 40 / 15 } -- updated 6/17
 game.TraitData.DoorHealToFullBoon.RarityLevels.Perfect = { Multiplier = 3.5 }
 game.TraitData.WeakPotencyBoon.RarityLevels.Perfect = { Multiplier = 2.0 } -- added 2/19
 game.TraitData.WeakVulnerabilityBoon.RarityLevels.Perfect = { Multiplier = 3.5 }
@@ -20,7 +20,8 @@ game.TraitData.FocusRawDamageBoon.RarityLevels.Perfect = { Multiplier = 15 / 5 }
 game.TraitData.ApolloWeaponBoon.RarityLevels.Perfect = { Multiplier = 3.5 }
 game.TraitData.ApolloSpecialBoon.RarityLevels.Perfect = { Multiplier = 2.67 }
 game.TraitData.ApolloCastBoon.RarityLevels.Perfect = { Multiplier = 3.5 }
-game.TraitData.ApolloExCastBoon.RarityLevels.Perfect = { Multiplier = 2.0 }
+game.TraitData.ApolloExCastBoon.RarityLevels.Heroic = { Multiplier = 2.5 } -- updated 6/17, temp bugfix
+game.TraitData.ApolloExCastBoon.RarityLevels.Perfect = { Multiplier = 3.5 } -- updated 6/17
 game.TraitData.ApolloSprintBoon.RarityLevels.Perfect = { Multiplier = 8 / 3 } -- 1/3
 game.TraitData.ApolloManaBoon.RarityLevels.Perfect = { Multiplier = 3.5 }
 game.TraitData.ApolloRetaliateBoon.RarityLevels.Perfect = { Multiplier = 3.5 }
@@ -34,27 +35,28 @@ game.TraitData.DoubleStrikeChanceBoon.RarityLevels.Perfect = { Multiplier = 3.5 
 -- ares added 2/19
 game.TraitData.AresWeaponBoon.RarityLevels.Perfect = { Multiplier = 70 / 20 }
 game.TraitData.AresSpecialBoon.RarityLevels.Perfect = { Multiplier = 80 / 30 }
-game.TraitData.AresManaBoon.RarityLevels.Perfect = { Multiplier = 2.25 }
+game.TraitData.AresManaBoon.RarityLevels.Perfect = { Multiplier = 2 } -- updated 6/17
 game.TraitData.AresCastBoon.RarityLevels.Perfect = { Multiplier = 230 / 80 }
 game.TraitData.AresSprintBoon.RarityLevels.Perfect = { Multiplier = 3.5 }
 game.TraitData.AresExCastBoon.RarityLevels.Perfect = { Multiplier = 80 / 30 }
-game.TraitData.RendBloodDropBoon.RarityLevels.Perfect = { Multiplier = 35 / 10 }
-game.TraitData.AresStatusDoubleDamageBoon.RarityLevels.Perfect = { Multiplier = 35 / 10 }
-game.TraitData.BloodDropRevengeBoon.RarityLevels.Perfect = { Multiplier = 75 / 25 }
+game.TraitData.RendBloodDropBoon.RarityLevels.Perfect = { Multiplier = 2 } -- updated 6/17
+game.TraitData.AresStatusDoubleDamageBoon.RarityLevels.Perfect = { Multiplier = 2 } -- updated 6/17
+game.TraitData.BloodDropRevengeBoon.RarityLevels.Perfect = { Multiplier = 2 } -- updated 6/17
 game.TraitData.MissingHealthCritBoon.RarityLevels.Perfect = { Multiplier = 0.2 / 0.1 }
-game.TraitData.AloneDamageBoon.RarityLevels.Perfect = { Multiplier = 45 / 20 }
+--game.TraitData.AloneDamageBoon.RarityLevels.Perfect = { Multiplier = 45 / 20 } removed from ares 6/17
 game.TraitData.LowHealthLifestealBoon.RarityLevels.Perfect = { Multiplier = 6 }
 game.TraitData.OmegaDelayedDamageBoon.RarityLevels.Perfect = { Multiplier = 240 / 90 }
 
 -- artemis
 game.TraitData.InsideCastCritBoon.RarityLevels.Perfect = { Multiplier = 2.5 }
-game.TraitData.OmegaCastVolleyBoon.RarityLevels.Perfect = { Multiplier = 80 / 30 } -- 10/30
+game.TraitData.OmegaCastVolleyBoon.RarityLevels.Perfect = { Multiplier = 2.25 } -- update 6/17
 game.TraitData.HighHealthCritBoon.RarityLevels.Perfect = { Multiplier = 3.5 }
 game.TraitData.CritBonusBoon.RarityLevels.Perfect = { Multiplier = 2.67 }
 game.TraitData.DashOmegaBuffBoon.RarityLevels.Perfect = { Multiplier = 2.25 }
 game.TraitData.SupportingFireBoon.RarityLevels.Perfect = { Multiplier = 3.5 }
 game.TraitData.TimedCritVulnerabilityBoon.RarityLevels.Perfect = { Multiplier = 10 / 20 } -- -2/20
-game.TraitData.FocusCritBoon.RarityLevels.Perfect = { Multiplier = 2 } -- added 2/19
+game.TraitData.FocusCritBoon.RarityLevels.Perfect = { Multiplier = 2 }                    -- added 2/19
+game.TraitData.SorceryCritBoon.RarityLevels.Perfect = { Multiplier = 0.55 / 0.3 } -- added 6/17
 
 -- athena
 game.TraitData.InvulnerabilityDashBoon.RarityLevels.Perfect = { Multiplier = 3.5 }
@@ -64,6 +66,7 @@ game.TraitData.AthenaProjectileBoon.RarityLevels.Perfect = { Multiplier = 3.5 }
 game.TraitData.DeathDefianceRefillBoon.RarityLevels.Perfect = { Multiplier = 3.5 }
 game.TraitData.InvulnerabilityCastBoon.RarityLevels.Perfect = { Multiplier = 3.5 }
 game.TraitData.ManaSpearBoon.RarityLevels.Perfect = { Multiplier = 400 / 150 } -- added 2/19
+game.TraitData.OlympianSpellCountBoon.RarityLevels.Perfect = { Multiplier = 6 } -- added 6/17
 
 -- demeter
 game.TraitData.DemeterWeaponBoon.RarityLevels.Perfect = { Multiplier = 3.5 }
@@ -74,22 +77,24 @@ game.TraitData.DemeterManaBoon.RarityLevels.Perfect = { Multiplier = 3.5 }
 game.TraitData.CastNovaBoon.RarityLevels.Perfect = { Multiplier = 6 }          -- 1
 game.TraitData.PlantHealthBoon.RarityLevels.Perfect = { Multiplier = 2.0 } -- updated 2/19
 game.TraitData.BoonGrowthBoon.RarityLevels.Perfect = { Multiplier = 1/6 } -- updated 2/19
-game.TraitData.ReserveManaHitShieldBoon.RarityLevels.Perfect = { Multiplier = 0.01/20 } -- -5/20, updated 2/19
-game.TraitData.SlowExAttackBoon.RarityLevels.Perfect = { Multiplier = 3.5 }
+game.TraitData.ReserveManaHitShieldBoon.RarityLevels.Perfect = { Multiplier = 1 / 25 } -- updated 6/17, best i can come up wtih
+game.TraitData.SlowExAttackBoon.RarityLevels.Perfect = { Multiplier = 2.5 } -- updated 6/17
 game.TraitData.CastAttachBoon.RarityLevels.Perfect = { Multiplier = 3.5 }
 game.TraitData.RootDurationBoon.RarityLevels.Perfect = { Multiplier = 3.5 }
 
 -- dionysus
 game.TraitData.CastLobBoon.RarityLevels.Perfect = { Multiplier = 3.5 }
-game.TraitData.HiddenMaxHealthBoon.RarityLevels.Perfect = { Multiplier = 2.25 }
+game.TraitData.HiddenMaxHealthBoon.RarityLevels.Perfect = { Multiplier = 3 } -- updated 6/17
 game.TraitData.FirstHangoverBoon.RarityLevels.Perfect = { Multiplier = 3.5 }
 game.TraitData.PowerDrinkBoon.RarityLevels.Perfect = { Multiplier = 5 / 10 }
 game.TraitData.CombatEncounterHealBoon.RarityLevels.Perfect = { Multiplier = 2.0 }
 game.TraitData.FogDamageBonusBoon.RarityLevels.Perfect = { Multiplier = 2.67 }
+game.TraitData.BankBoon.RarityLevels.Perfect = { Multiplier = 800 / 300 } -- updated 6/17
+game.TraitData.RandomBaseDamageBoon.RarityLevels.Perfect = { Multiplier = 3 } -- added 6/17
 
 -- hephaestus
 game.TraitData.HephaestusWeaponBoon.RarityLevels.Perfect = { Multiplier = 2 / 12 }   -- -2/12
-game.TraitData.HephaestusSpecialBoon.RarityLevels.Perfect = { Multiplier = 5 / 16 } -- updated 2/19
+game.TraitData.HephaestusSpecialBoon.RarityLevels.Perfect = { Multiplier = 5 / 15 } -- updated 6/17
 game.TraitData.HephaestusCastBoon.RarityLevels.Perfect = { Multiplier = 2.2 }        -- 0.2
 game.TraitData.HephaestusSprintBoon.RarityLevels.Perfect = { Multiplier = 5 / 10 } -- changed 10/16
 game.TraitData.HephaestusManaBoon.RarityLevels.Perfect = { Multiplier = 6 }
@@ -97,10 +102,10 @@ game.TraitData.HephaestusManaBoon.RarityLevels.Perfect = { Multiplier = 6 }
 game.TraitData.AntiArmorBoon.RarityLevels.Perfect = { Multiplier = 3.5 }
 game.TraitData.HeavyArmorBoon.RarityLevels.Perfect = { Multiplier = 3.5 }
 game.TraitData.ArmorBoon.RarityLevels.Perfect = { Multiplier = 3.5 }
-game.TraitData.EncounterStartDefenseBuffBoon.RarityLevels.Perfect = { Multiplier = 11 / 6 }
-game.TraitData.ManaToHealthBoon.RarityLevels.Perfect = { Multiplier = 3.5 }
-game.TraitData.MassiveDamageBoon.RarityLevels.Perfect = { Multiplier = 3.5 }
-game.TraitData.MassiveKnockupBoon.RarityLevels.Perfect = { Multiplier = 8 / 3 } -- 1/3
+game.TraitData.EncounterStartDefenseBuffBoon.RarityLevels.Perfect = { Multiplier = 200 / 75 } -- updated 6/17
+game.TraitData.ManaToHealthBoon.RarityLevels.Perfect = { Multiplier = 2.25 } -- updated 6/17
+game.TraitData.MassiveDamageBoon.RarityLevels.Perfect = { Multiplier = 200 / 75 } -- updated 6/17
+game.TraitData.MassiveKnockupBoon.RarityLevels.Perfect = { Multiplier = 1.125 } -- updated 6/17
 
 -- hera
 game.TraitData.HeraWeaponBoon.RarityLevels.Perfect = { Multiplier = 2 }
@@ -110,10 +115,10 @@ game.TraitData.SpawnCastDamageBoon.RarityLevels.Perfect = { Multiplier = 8 / 3 }
 game.TraitData.HeraSprintBoon.RarityLevels.Perfect = { Multiplier = 3.5 }
 game.TraitData.HeraManaBoon.RarityLevels.Perfect = { Multiplier = 0.5 } -- -0.1
 game.TraitData.CommonGlobalDamageBoon.RarityLevels.Perfect = { Multiplier = 3.5 }
-game.TraitData.DamageShareRetaliateBoon.RarityLevels.Perfect = { Multiplier = 14 / 4 }
+game.TraitData.DamageShareRetaliateBoon.RarityLevels.Perfect = { Multiplier = 8 / 3 } -- updated 6/17
 --game.TraitData.SwapBonusBoon.RarityLevels.Perfect = { Multiplier = 3.5 } removed? 10/16
 game.TraitData.BoonDecayBoon.RarityLevels.Perfect = { Multiplier = 6 } -- updated 2/19
-game.TraitData.OmegaHeraProjectileBoon.RarityLevels.Perfect = { Multiplier = 2.0 }
+game.TraitData.OmegaHeraProjectileBoon.RarityLevels.Perfect = { Multiplier = 2.25 } -- updated 6/17
 game.TraitData.DamageSharePotencyBoon.RarityLevels.Perfect = { Multiplier = 3.5 }
 game.TraitData.LinkedDeathDamageBoon.RarityLevels.Perfect = { Multiplier = 3.5 }
 --game.TraitData.FullManaExBoostBoon.RarityLevels.Perfect = { Multiplier = 3.5 } boon removed 6/26
@@ -134,12 +139,13 @@ game.TraitData.RestockBoon.RarityLevels.Perfect = { Multiplier = 3.5 } -- added 
 game.TraitData.LuckyBoon.RarityLevels.Perfect = { Multiplier = 80 / 30 } -- added 2/19
 
 -- hestia
-game.TraitData.HestiaWeaponBoon.RarityLevels.Perfect = { Multiplier = 2.67 } -- 0.33ish
-game.TraitData.HestiaSpecialBoon.RarityLevels.Perfect = { Multiplier = 2.25 }
+game.TraitData.HestiaWeaponBoon.RarityLevels.Perfect = { Multiplier = 3.5 } -- updated 6/17
+game.TraitData.HestiaSpecialBoon.RarityLevels.Perfect = { Multiplier = 3.5 } -- updated 6/17
 game.TraitData.HestiaCastBoon.RarityLevels.Perfect = { Multiplier = 3.5 }
-game.TraitData.HestiaSprintBoon.RarityLevels.Perfect = { Multiplier = 6 }
+game.TraitData.HestiaSprintBoon.RarityLevels.Perfect = { Multiplier = 3.5 } -- updated 6/17
 game.TraitData.HestiaManaBoon.RarityLevels.Perfect = { Multiplier = 3.5 }
-game.TraitData.SacrificeBoon.RarityLevels.Perfect = { Multiplier = 6 } -- updated 2/19
+game.TraitData.AloneDamageBoon.RarityLevels.Perfect = { Multiplier = 2 } -- added to hestia 6/17
+--game.TraitData.SacrificeBoon.RarityLevels.Perfect = { Multiplier = 6 } -- removed 6/17
 game.TraitData.OmegaZeroBurnBoon.RarityLevels.Perfect = { Multiplier = 3.5 }
 game.TraitData.CastProjectileBoon.RarityLevels.Perfect = { Multiplier = 3.5 }
 game.TraitData.FireballManaSpecialBoon.RarityLevels.Perfect = { Multiplier = 3.5 }
@@ -149,31 +155,31 @@ game.TraitData.BurnArmorBoon.RarityLevels.Perfect = { Multiplier = 2.25 } -- upd
 game.TraitData.BurnStackBoon.RarityLevels.Perfect = { Multiplier = 3.5 }
 
 -- poseidon
-game.TraitData.PoseidonWeaponBoon.RarityLevels.Perfect = { Multiplier = 2.67 }
-game.TraitData.PoseidonSpecialBoon.RarityLevels.Perfect = { Multiplier = 3.5 }
+game.TraitData.PoseidonWeaponBoon.RarityLevels.Perfect = { Multiplier = 2.25 } -- updated 6/17
+game.TraitData.PoseidonSpecialBoon.RarityLevels.Perfect = { Multiplier = 2.67 } -- updated 6/17
 game.TraitData.PoseidonCastBoon.RarityLevels.Perfect = { Multiplier = 3.5 }
 game.TraitData.PoseidonExCastBoon.RarityLevels.Perfect = { Multiplier = 400 / 150 } -- updated 2/19
-game.TraitData.OmegaPoseidonProjectileBoon.RarityLevels.Perfect = { Multiplier = 2.0 }
-game.TraitData.PoseidonSprintBoon.RarityLevels.Perfect = { Multiplier = 3.5 }
-game.TraitData.PoseidonManaBoon.RarityLevels.Perfect = { Multiplier = 2.0 }
-game.TraitData.EncounterStartOffenseBuffBoon.RarityLevels.Perfect = { Multiplier = 2.67 }
+game.TraitData.OmegaPoseidonProjectileBoon.RarityLevels.Perfect = { Multiplier = 3.5 } -- updated 6/17
+game.TraitData.PoseidonSprintBoon.RarityLevels.Perfect = { Multiplier = 180 / 80 } -- updated 6/17
+game.TraitData.PoseidonManaBoon.RarityLevels.Perfect = { Multiplier = 9 / 4 } -- updated 6/17
+game.TraitData.EncounterStartOffenseBuffBoon.RarityLevels.Perfect = { Multiplier = 2.25 } -- updated 6/17
 --game.TraitData.MinorLootBoon.RarityLevels.Perfect = { Multiplier = 2.5 } -- 0.3 boon removed 6/26
 game.TraitData.RoomRewardBonusBoon.RarityLevels.Perfect = { Multiplier = 3.5 }
 game.TraitData.FocusDamageShaveBoon.RarityLevels.Perfect = { Multiplier = 3.5 }
-game.TraitData.DoubleRewardBoon.RarityLevels.Perfect = { Multiplier = 2.25 }
-game.TraitData.PoseidonStatusBoon.RarityLevels.Perfect = { Multiplier = 6 } -- added 2/19
+game.TraitData.DoubleRewardBoon.RarityLevels.Perfect = { Multiplier = 2 } -- updated 6/17
+game.TraitData.PoseidonStatusBoon.RarityLevels.Perfect = { Multiplier = 2.25 } -- updated 6/17
 
 -- zeus
 game.TraitData.ZeusWeaponBoon.RarityLevels.Perfect = { Multiplier = 2.8 } -- 0.4
-game.TraitData.ZeusSpecialBoon.RarityLevels.Perfect = { Multiplier = 3 }
+game.TraitData.ZeusSpecialBoon.RarityLevels.Perfect = { Multiplier = 2.5 } -- upated 6/17
 game.TraitData.ZeusCastBoon.RarityLevels.Perfect = { Multiplier = 2.25 }
 -- game.TraitData.ZeusExCastBoon.RarityLevels.Perfect = { Multiplier = 3.5 } removed 2/19
 game.TraitData.ZeusSprintBoon.RarityLevels.Perfect = { Multiplier = 2.25 }
 game.TraitData.ZeusManaBoon.RarityLevels.Perfect = { Multiplier = 5 / 10 } -- -1/10
-game.TraitData.ZeusManaBoltBoon.RarityLevels.Perfect = { Multiplier = 2.25 }
+game.TraitData.ZeusManaBoltBoon.RarityLevels.Perfect = { Multiplier = 3.5 } -- updated 6/17
 -- note - BoltRetaliateBoon has no rarity
 game.TraitData.CastAnywhereBoon.RarityLevels.Perfect = { Multiplier = 3 }
 game.TraitData.FocusLightningBoon.RarityLevels.Perfect = { Multiplier = 3.5 }
 game.TraitData.DoubleBoltBoon.RarityLevels.Perfect = { Multiplier = 3.5 }
-game.TraitData.EchoExpirationBoon.RarityLevels.Perfect = { Multiplier = 45 / 20 } -- updated 2/19
-game.TraitData.LightningDebuffGeneratorBoon.RarityLevels.Perfect = { Multiplier = 3.5 }
+game.TraitData.EchoExpirationBoon.RarityLevels.Perfect = { Multiplier = 80 / 30 } -- updated 6/17
+game.TraitData.LightningDebuffGeneratorBoon.RarityLevels.Perfect = { Multiplier = 2.25 } -- updated 6/17